fix(#69) : pyodide 오류 해결
